+++ core/files/fileoffloader/offloader.go
@@ -76,7 +76,7 @@ func (s *service) fileOffload(ctx context.Context, fileDetails *domain.Details, 
 	if fileId == "" {
 		return 0, fmt.Errorf("empty file id")
 	}
-	backupStatus := filesyncstatus.Status(fileDetails.GetIntOrDefault(bundle.RelationKeyFileBackupStatus, 0))
+	backupStatus := filesyncstatus.Status(fileDetails.GetInt64OrDefault(bundle.RelationKeyFileBackupStatus, 0))
 	id := domain.FullFileId{
 		SpaceId: fileDetails.GetStringOrDefault(bundle.RelationKeySpaceId, ""),
 		FileId:  domain.FileId(fileId),
